requirements-docker.txt
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.1
fastapi==0.121.3
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
loguru==0.7.3
MarkupSafe==2.1.5
matplotlib==3.10.7
ml_dtypes==0.5.4
mpmath==1.3.0
narwhals==2.12.0
networkx==3.3
numpy==1.26.4
onnx==1.19.1
onnxruntime==1.23.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.1
pyarrow==21.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.29.0
scikit-learn==1.7.2
scipy==1.11.4
seaborn==0.13.2
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.50.0
streamlit==1.51.0
streamlit-drawable-canvas==0.9.3
stringzilla==4.2.3
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
win32_setctime==1.2.0
